Move grdinfo function to grdinfo.py
