+++ pygmt/src/__init__.py
@@ -5,6 +5,7 @@ Source code for PyGMT modules.
 from pygmt.src.blockmedian import blockmedian
 from pygmt.src.grdcut import grdcut
 from pygmt.src.grdfilter import grdfilter
+from pygmt.src.grdinfo import grdinfo
 from pygmt.src.grdtrack import grdtrack
 from pygmt.src.info import info
 from pygmt.src.inset import inset
